requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
tenacity
requests
pytest
pytest-asyncio
respx
